fix(web): decode breadcrumb path labels
